# website/data/openapi/dotmarc-api.json
{
  "components": {
    "schemas": {
...
      "ApiAddDomainRequest": {
        "example": {
          "name": "contoso.example"
        },
        "properties": {
          "name": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ApiCheck": {
        "example": {
          "checkedUtc": "2026-10-01T09:30:00+00:00",
          "detail": null,
          "status": "Ok"
        },
        "properties": {
          "checkedUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "checkedUtc",
          "detail"
        ],
        "type": "object"
      },
      "ApiDmarcPolicy": {
        "properties": {
          "percent": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "policy": {
            "nullable": true,
            "type": "string"
          },
          "subdomainPolicy": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "policy",
          "subdomainPolicy",
          "percent"
        ],
        "type": "object"
      },
      "ApiDnsProvider": {
        "properties": {
          "provider": {
            "type": "string"
          },
          "zone": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "provider",
          "zone"
        ],
        "type": "object"
      },
      "ApiDomain": {
        "example": {
          "groups": [
            {
              "id": 3,
              "name": "Contoso"
            }
          ],
          "id": 42,
          "lastReportReceivedUtc": "2026-10-01T09:30:00+00:00",
          "monitored": true,
          "name": "contoso.example",
          "passRate": 0.987,
          "tags": [
            {
              "id": 7,
              "name": "Microsoft 365"
            }
          ]
        },
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ApiNamedRef"
            },
            "type": "array"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "lastReportReceivedUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "monitored": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "passRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ApiNamedRef"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "monitored",
          "groups",
          "tags",
          "lastReportReceivedUtc",
          "passRate"
        ],
        "type": "object"
      },
      "ApiDomainDetail": {
        "properties": {
          "dmarcPolicy": {
            "$ref": "#/components/schemas/ApiDmarcPolicy"
          },
          "dnsProvider": {
            "$ref": "#/components/schemas/ApiDnsProvider"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ApiNamedRef"
            },
            "type": "array"
          },
          "health": {
            "$ref": "#/components/schemas/ApiDomainHealth"
          },
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "lastReportReceivedUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "monitored": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "passRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ApiNamedRef"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "monitored",
          "groups",
          "tags",
          "lastReportReceivedUtc",
          "passRate",
          "health",
          "dmarcPolicy",
          "dnsProvider"
        ],
        "type": "object"
      },
      "ApiDomainHealth": {
        "properties": {
          "dkim": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "dmarc": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "dmarcAuthorization": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "mtaSts": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "mx": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "spf": {
            "$ref": "#/components/schemas/ApiCheck"
          },
          "tlsrpt": {
            "$ref": "#/components/schemas/ApiCheck"
          }
        },
        "required": [
          "dmarc",
          "spf",
          "dkim",
          "mx",
          "tlsrpt",
          "mtaSts",
          "dmarcAuthorization"
        ],
        "type": "object"
      },
...
      "ApiNamedRef": {
        "properties": {
          "id": {
            "format": "int32",
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
...
      "ApiPageOfApiDomain": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiDomain"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "pageSize": {
            "format": "int32",
            "type": "integer"
          },
          "totalCount": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object"
      },
      "ApiReasonBreakdown": {
        "properties": {
          "benignOverride": {
            "format": "int32",
            "type": "integer"
          },
          "inferredBothFailure": {
            "format": "int32",
            "type": "integer"
          },
          "inferredDkimFailure": {
            "format": "int32",
            "type": "integer"
          },
          "inferredSpfFailure": {
            "format": "int32",
            "type": "integer"
          },
          "localPolicy": {
            "format": "int32",
            "type": "integer"
          },
          "noReasonGiven": {
            "format": "int32",
            "type": "integer"
          },
          "other": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "benignOverride",
          "localPolicy",
          "other",
          "noReasonGiven",
          "inferredSpfFailure",
          "inferredDkimFailure",
          "inferredBothFailure"
        ],
        "type": "object"
      },
      "ApiReportSummary": {
        "properties": {
          "days": {
            "format": "int32",
            "type": "integer"
          },
          "domainId": {
            "format": "int32",
            "type": "integer"
          },
          "domainName": {
            "type": "string"
          },
          "passRate": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "reasonBreakdown": {
            "$ref": "#/components/schemas/ApiReasonBreakdown"
          },
          "topSources": {
            "items": {
              "$ref": "#/components/schemas/ApiSource"
            },
            "type": "array"
          },
          "totalVolume": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "domainId",
          "domainName",
          "days",
          "totalVolume",
          "passRate",
          "reasonBreakdown",
          "topSources"
        ],
        "type": "object"
      },
...
      "ApiSource": {
        "example": {
          "disposition": "None",
          "dkim": "Pass",
          "sourceIp": "203.0.113.25",
          "spf": "Pass",
          "volume": 1480
        },
        "properties": {
          "disposition": {
            "type": "string"
          },
          "dkim": {
            "type": "string"
          },
          "sourceIp": {
            "type": "string"
          },
          "spf": {
            "type": "string"
          },
          "volume": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "sourceIp",
          "volume",
          "spf",
          "dkim",
          "disposition"
        ],
        "type": "object"
      },
...
      "HttpValidationProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ProblemDetails": {
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "dmk_ API key",
        "description": "An API key created on dotMARC's Access page.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read domains, DNS health, DMARC report summaries, groups, tags and alerts, and add, import and organise domains. Authenticate with an API key from dotMARC's Access page: `Authorization: Bearer dmk_...`. Each operation needs the permission named in its description, and a key limited to certain groups only sees their domains. Each key may make 120 requests a minute.",
    "title": "dotMARC API",
    "version": "0.9.0"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/v1/domains": {
      "get": {
        "description": "Domains this key can see, in dashboard order, 50 a page by default and at most 200. Filter by group id, tag id or monitoring. The pass rate covers the last 30 days and is null with no reports.\n\nNeeds the DomainsView permission.",
        "operationId": "ListDomains",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "group",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "monitored",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPageOfApiDomain"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "The request couldn't be used. `errors` names each field that's wrong and why."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "List domains",
        "tags": [
          "Domains"
        ],
        "x-dotmarc-permission": "DomainsView"
      },
      "post": {
        "description": "Adds a domain for dotMARC to monitor. It shows as missing reports until its first DMARC report arrives.\n\nNeeds the DomainsAdd permission.",
        "operationId": "AddDomain",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiAddDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDomain"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "The request couldn't be used. `errors` names each field that's wrong and why."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with how things are now, such as a domain that's already in dotMARC or an alert that can't be acknowledged. `detail` says what."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "Add a domain",
        "tags": [
          "Domains"
        ],
        "x-dotmarc-permission": "DomainsAdd"
      }
    },
...
    "/api/v1/domains/{domain}": {
      "get": {
        "description": "The domain with the results of dotMARC's last DNS checks. The API never runs checks itself; checkedUtc says when each last ran. Accepts the domain's id or its name, such as contoso.example.\n\nNeeds the DomainsView permission.",
        "operationId": "GetDomain",
        "parameters": [
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiDomainDetail"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "There's no such item, or this key can't see it. A key limited to certain groups can't see what's outside them."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "Get a domain and its health",
        "tags": [
          "Domains"
        ],
        "x-dotmarc-permission": "DomainsView"
      }
    },
...
    "/api/v1/domains/{domain}/reports/summary": {
      "get": {
        "description": "Volume, pass rate, why failing mail was let through or rejected, and the 20 busiest sending IPs, over the last 1 to 30 days (default 30). Accepts the domain's id or its name, such as contoso.example.\n\nNeeds the DomainsView permission.",
        "operationId": "GetReportSummary",
        "parameters": [
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiReportSummary"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            },
            "description": "The request couldn't be used. `errors` names each field that's wrong and why."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "No API key was sent, or the key is unknown, expired or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The key's role doesn't have the permission this operation needs, or the key is limited to certain groups and this would reach outside them. `detail` says which."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "There's no such item, or this key can't see it. A key limited to certain groups can't see what's outside them."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "This key has made more requests this minute than dotMARC allows (120 by default). Wait for the number of seconds in the Retry-After header, then try again."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Something went wrong in dotMARC that the request didn't cause. It's safe to retry a read; check before retrying a change. If it keeps happening, dotMARC's logs have the details."
          }
        },
        "summary": "Summarise a domain's DMARC reports",
        "tags": [
          "Domains"
        ],
        "x-dotmarc-permission": "DomainsView"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}